// 02-my-dashboard/src/pokemons/interfaces/pokemon.ts
export interface PokemonMove {
    move: {
        name: string;
        url: string;
    };
    version_group_details: {
        level_learned_at: number;
        move_learn_method: {
            name: string;
            url: string;
        };
        version_group: {
            name: string;
            url: string;
        };
    }[];
}

export interface PokemonStat {
    base_stat: number;
    effort: number;
    stat: {
        name: string;
        url: string;
    };
}

export interface EvolutionMoveData {
    name: string;
    id: string;
    moves: PokemonMove[];
}

export interface EvolutionMoveDisplay {
    name: string;
    id: string;
    moves: string[];
}

export interface AbilityDisplay {
    name: string;
    isHidden: boolean;
}

export interface PokemonPageProps {
    params: {
        name: string;
        id: string;
    };
}